add vpnd watch
watch connection state with the grpc server
