+++ nym-vpn-desktop/src-tauri/src/commands/mod.rs
@@ -1,6 +1,7 @@
 pub mod cli;
 pub mod connection;
 pub mod credential;
+pub mod daemon;
 pub mod db;
 pub mod log;
 pub mod node_location;
